ai-dm-voice/requirements.txt
# Core
flask>=3.0.0
python-dotenv>=1.0.0

# Discord
discord.py>=2.3.0
PyNaCl>=1.5.0  # Required for voice support

# AI Services
openai>=1.0.0
requests>=2.31.0

# TTS
edge-tts>=6.1.0  # Free Microsoft TTS fallback
